restructure into managers and models
