+++ Bar Tasker/BarTaskerCoordinator.swift
@@ -1,42 +1,70 @@
 import AppKit
-import Combine
 import Foundation
+import Observation
 import OSLog
 import ServiceManagement
 import SwiftUI
 
 @MainActor
-class BarTaskerManager: ObservableObject {
-  let logger = Logger(subsystem: "uk.co.maybeitsadam.bar-tasker", category: "manager")
+@Observable class BarTaskerCoordinator {
+  @ObservationIgnored let logger = Logger(subsystem: "uk.co.maybeitsadam.bar-tasker", category: "manager")
 
-  @Published var username: String
-  @Published var remoteKey: String
-  @Published var listId: String
+  // MARK: - Task Repository
 
-  /// All tasks (flat, from API)
-  @Published var tasks: [CheckvistTask] = []
+  let repository: TaskRepository
 
-  /// User's available lists from the API
-  @Published var availableLists: [CheckvistList] = []
+  // MARK: - Forwarding properties (heavily used across extensions + views)
 
-  /// The parent ID of the level currently being viewed (0 = root)
-  @Published var currentParentId: Int = 0
+  var tasks: [CheckvistTask] {
+    get { repository.tasks }
+    set { repository.tasks = newValue }
+  }
+  var currentParentId: Int {
+    get { repository.currentParentId }
+    set { repository.currentParentId = newValue }
+  }
+  var currentSiblingIndex: Int {
+    get { repository.currentSiblingIndex }
+    set { repository.currentSiblingIndex = newValue }
+  }
+  var listId: String {
+    get { repository.listId }
+    set { repository.listId = newValue }
+  }
+  var errorMessage: String? {
+    get { repository.errorMessage }
+    set { repository.errorMessage = newValue }
+  }
+  var lastUndo: UndoableAction? {
+    get { repository.lastUndo }
+    set { repository.lastUndo = newValue }
+  }
 
-  /// Index within the current level's sibling list
-  @Published var currentSiblingIndex: Int = 0
+  // MARK: - View / Filter State
 
-  @Published var isLoading: Bool = false
-  @Published var errorMessage: String? = nil
-
-  // MARK: - Undo
-  @Published var lastUndo: UndoableAction? = nil
-
-  @Published var hideFuture: Bool = false
-  @Published var rootTaskView: RootTaskView
-  @Published var selectedRootDueBucketRawValue: Int
-  @Published var selectedRootTag: String
+  var hideFuture: Bool = false {
+    didSet { invalidateCaches() }
+  }
+  var rootTaskView: RootTaskView {
+    didSet {
+      preferencesStore.set(rootTaskView.rawValue, for: .rootTaskView)
+      invalidateCaches()
+    }
+  }
+  var selectedRootDueBucketRawValue: Int {
+    didSet {
+      preferencesStore.set(selectedRootDueBucketRawValue, for: .selectedRootDueBucketRawValue)
+      invalidateCaches()
+    }
+  }
+  var selectedRootTag: String {
+    didSet {
+      preferencesStore.set(selectedRootTag, for: .selectedRootTag)
+      invalidateCaches()
+    }
+  }
   /// 0 = task list, 1 = root tabs (All/Due/Tags), 2 = root filter row (due buckets/tags)
-  @Published var rootScopeFocusLevel: Int = 0
+  var rootScopeFocusLevel: Int = 0
   var orderedRootTaskViews: [RootTaskView] {
     if let data = UserDefaults.standard.data(forKey: "rootTaskViewOrder"),
        let rawValues = try? JSONDecoder().decode([Int].self, from: data) {
@@ -49,16 +77,13 @@ class BarTaskerManager: ObservableObject {
     }
     return RootTaskView.allCases
   }
-  
+
   func saveRootTaskViewOrder(_ views: [RootTaskView]) {
     let rawValues = views.map { $0.rawValue }
     if let data = try? JSONEncoder().encode(rawValues) {
       UserDefaults.standard.set(data, forKey: "rootTaskViewOrder")
     }
   }
-  @Published var priorityTaskIds: [Int]
-
-  static let maxPriorityRank = 9
 
   // MARK: - Carbon hotkey constants
   enum CarbonKey {
@@ -69,7 +94,6 @@ class BarTaskerManager: ObservableObject {
     static let option = 0x0800
     static let shiftOption = 0x0A00
   }
-  private static let priorityQueuesDefaultsKey = "priorityTaskIdsByListId"
 
   // MARK: - Start Dates
   let startDates: StartDateManager
@@ -81,47 +105,31 @@ class BarTaskerManager: ObservableObject {
   let timer: TimerManager
 
   // MARK: - Integrations
-  @Published var integrations: IntegrationCoordinator
+  var integrations: IntegrationCoordinator
 
   // MARK: - Quick Entry
-  @Published var quickEntry: QuickEntryManager
+  var quickEntry: QuickEntryManager
 
   // MARK: - Kanban
   let kanban: KanbanManager
 
   let preferences: PreferencesManager
-  @Published var onboardingCompleted: Bool
-  @Published var activeOnboardingDialog: OnboardingDialog?
-  @Published var isNetworkReachable: Bool = true
+  var onboardingCompleted: Bool {
+    didSet { preferencesStore.set(onboardingCompleted, for: .onboardingCompleted) }
+  }
+  var activeOnboardingDialog: OnboardingDialog?
 
-  /// Mutations that failed due to network unavailability, pending a retry when connectivity is restored.
-  /// Keyed by taskId; later mutations overwrite earlier ones for the same task.
-  var pendingTaskMutations: [Int: (content: String?, due: String?)] = [:]
-
-  var dismissedOnboardingDialogs: Set<OnboardingDialog>
-  var offlineArchivedTasksById: [Int: CheckvistTask]
-  var nextOfflineTaskIdValue: Int
+  @ObservationIgnored var dismissedOnboardingDialogs: Set<OnboardingDialog>
 
   // MARK: - Computed property caches
 
-  var cache = BarTaskerCacheState()
+  @ObservationIgnored var cache = BarTaskerCacheState()
 
-  var loadingOperationCount: Int = 0
-  var cancellables = Set<AnyCancellable>()
-  let reorderQueue = BarTaskerReorderQueue()
-  var isApplyingLaunchAtLoginChange = false
-  var hasAttemptedRemoteKeyBootstrap = false
-  let preferencesStore = BarTaskerPreferencesStore()
-  let localTaskStore: LocalTaskStore
-  let checkvistSyncPlugin: any CheckvistSyncPlugin
+  @ObservationIgnored var isApplyingLaunchAtLoginChange = false
+  @ObservationIgnored let preferencesStore = BarTaskerPreferencesStore()
   let userPluginManager: UserPluginManager
-  lazy var commandExecutor = BarTaskerCommandExecutor(manager: self)
-  let navigationCoordinator = TaskNavigationCoordinator()
-  let reachabilityMonitor = NetworkReachabilityMonitor()
-  let priorityQueueStore = ListScopedTaskIDStore(
-    defaultsKey: BarTaskerManager.priorityQueuesDefaultsKey,
-    maximumCount: BarTaskerManager.maxPriorityRank
-  )
+  @ObservationIgnored lazy var commandExecutor = BarTaskerCommandExecutor(manager: self)
+  @ObservationIgnored let reachabilityMonitor = NetworkReachabilityMonitor()
   var usesKeychainStorage: Bool {
     #if DEBUG
       !preferences.ignoreKeychainInDebug
@@ -134,10 +142,26 @@ class BarTaskerManager: ObservableObject {
     CheckvistCredentials(username: username, remoteKey: remoteKey)
   }
 
+  var username: String {
+    get { repository.username }
+    set { repository.username = newValue }
+  }
+  var remoteKey: String {
+    get { repository.remoteKey }
+    set { repository.remoteKey = newValue }
+  }
+  var availableLists: [CheckvistList] {
+    get { repository.availableLists }
+    set { repository.availableLists = newValue }
+  }
+  var isLoading: Bool {
+    get { repository.isLoading }
+    set { repository.isLoading = newValue }
+  }
+
   // swiftlint:disable function_body_length
   init(pluginRegistry: BarTaskerPluginRegistry) {
     let resolvedLocalTaskStore = LocalTaskStore()
-    let offlinePayload = resolvedLocalTaskStore.load()
     let resolvedCheckvistSyncPlugin =
       pluginRegistry.activeCheckvistSyncPlugin ?? NativeCheckvistSyncPlugin()
     let resolvedObsidianPlugin =
@@ -149,8 +173,7 @@ class BarTaskerManager: ObservableObject {
     let resolvedMCPIntegrationPlugin =
       pluginRegistry.activeMCPIntegrationPlugin
       ?? NativeMCPIntegrationPlugin()
-    self.checkvistSyncPlugin = resolvedCheckvistSyncPlugin
-    self.localTaskStore = resolvedLocalTaskStore
+
     self.userPluginManager = UserPluginManager(
       builtInPluginIdentifiers: [
         resolvedCheckvistSyncPlugin.pluginIdentifier,
@@ -161,20 +184,32 @@ class BarTaskerManager: ObservableObject {
     )
     self.preferences = PreferencesManager(preferencesStore: preferencesStore)
 
-    let storedUsername = preferencesStore.string(.checkvistUsername)
+    // Compute initial remote key before creating repository
+    let useKeychainStorageAtInit: Bool
+    #if DEBUG
+      let ignoreAtInit = preferencesStore.bool(.ignoreKeychainInDebug, default: true)
+      useKeychainStorageAtInit = !ignoreAtInit
+    #else
+      useKeychainStorageAtInit = true
+    #endif
+
+    let initialRemoteKey = resolvedCheckvistSyncPlugin.startupRemoteKey(
+      useKeychainStorageAtInit: useKeychainStorageAtInit)
+
+    // Create task repository with all task-related state
+    self.repository = TaskRepository(
+      preferencesStore: preferencesStore,
+      checkvistSyncPlugin: resolvedCheckvistSyncPlugin,
+      localTaskStore: resolvedLocalTaskStore,
+      initialRemoteKey: initialRemoteKey
+    )
+
     let storedListId = preferencesStore.string(.checkvistListId)
+    let storedUsername = preferencesStore.string(.checkvistUsername)
     let storedOnboardingCompletedFlag = preferencesStore.optionalBool(.onboardingCompleted)
     let storedPluginSelectionOnboardingCompletedFlag = preferencesStore.optionalBool(
       .pluginSelectionOnboardingCompleted)
-    self.username = storedUsername
-    self.listId = storedListId
-    self.tasks =
-      storedListId.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty
-      ? offlinePayload.openTasks : []
-    self.offlineArchivedTasksById = Dictionary(
-      uniqueKeysWithValues: offlinePayload.archivedTasks.map { ($0.id, $0) })
-    self.nextOfflineTaskIdValue = max(offlinePayload.nextTaskId, 1)
-    self.priorityTaskIds = priorityQueueStore.load(for: storedListId)
+
     self.rootTaskView =
       RootTaskView(rawValue: preferencesStore.int(.rootTaskView, default: 1)) ?? .due
     self.selectedRootDueBucketRawValue = preferencesStore.int(
@@ -221,25 +256,8 @@ class BarTaskerManager: ObservableObject {
       persistedDismissedDialogs.compactMap(OnboardingDialog.init(rawValue:))
     )
 
-    let useKeychainStorageAtInit: Bool
-    #if DEBUG
-      let ignoreAtInit = preferencesStore.bool(.ignoreKeychainInDebug, default: true)
-      useKeychainStorageAtInit = !ignoreAtInit
-    #else
-      useKeychainStorageAtInit = true
-    #endif
-
-    self.remoteKey = resolvedCheckvistSyncPlugin.startupRemoteKey(
-      useKeychainStorageAtInit: useKeychainStorageAtInit)
-
     setupBindings()
-    for child: any ObservableObject in [
-      preferences, timer, kanban, startDates, recurrence, quickEntry, integrations,
-    ] {
-      child.objectWillChange
-        .sink { [weak self] _ in self?.objectWillChange.send() }
-        .store(in: &cancellables)
-    }
+    setupChildCallbacks()
     kanban.dataSource = self
     integrations.dataSource = self
     integrations.onIntegrationStateChanged = { [weak self] in
@@ -263,20 +281,20 @@ class BarTaskerManager: ObservableObject {
 
 // MARK: - KanbanTaskDataSource conformance
 
-extension BarTaskerManager: KanbanTaskDataSource {}
+extension BarTaskerCoordinator: KanbanTaskDataSource {}
 
 // MARK: - IntegrationDataSource conformance
 
-extension BarTaskerManager: IntegrationDataSource {}
+extension BarTaskerCoordinator: IntegrationDataSource {}
 
 // MARK: - Kanban move orchestration (cross-cutting: kanban + task CRUD)
 
-extension BarTaskerManager {
+extension BarTaskerCoordinator {
   @MainActor func moveCurrentTaskToKanbanColumn(direction: Int) async {
     guard let outcome = kanban.computeMoveCurrentTask(direction: direction) else { return }
     switch outcome {
     case .error(let msg):
-      errorMessage = msg
+      repository.errorMessage = msg
     case .update(let task, let newContent, let newDue):
       await updateTask(task: task, content: newContent, due: newDue)
     }
@@ -286,7 +304,7 @@ extension BarTaskerManager {
     guard let outcome = kanban.computeMoveTask(id: taskId, toColumn: targetColumn) else { return }
     switch outcome {
     case .error(let msg):
-      errorMessage = msg
+      repository.errorMessage = msg
     case .update(let task, let newContent, let newDue):
       await updateTask(task: task, content: newContent, due: newDue)
     }
