+++ Bar Tasker/Managers/IntegrationCoordinator.swift
@@ -1,6 +1,6 @@
 import AppKit
-import Combine
 import Foundation
+import Observation
 import OSLog
 
 /// Provides read-only access to task/list data that IntegrationCoordinator needs.
@@ -13,29 +13,45 @@ protocol IntegrationDataSource: AnyObject {
 }
 
 @MainActor
-class IntegrationCoordinator: ObservableObject {
-  private let logger = Logger(
+@Observable class IntegrationCoordinator {
+  @ObservationIgnored private let logger = Logger(
     subsystem: "uk.co.maybeitsadam.bar-tasker", category: "integrations")
-  private let preferencesStore: BarTaskerPreferencesStore
-  private var cancellables = Set<AnyCancellable>()
+  @ObservationIgnored private let preferencesStore: BarTaskerPreferencesStore
 
-  weak var dataSource: IntegrationDataSource?
+  @ObservationIgnored weak var dataSource: IntegrationDataSource?
 
   /// Called when integration-enabled flags change so the coordinator can refresh onboarding.
-  var onIntegrationStateChanged: (() -> Void)?
+  @ObservationIgnored var onIntegrationStateChanged: (() -> Void)?
+  /// Called with an error message (or nil to clear) when integration actions produce errors.
+  @ObservationIgnored var onError: ((String?) -> Void)?
 
   // MARK: - Integration enable flags
 
-  @Published var obsidianIntegrationEnabled: Bool
-  @Published var googleCalendarIntegrationEnabled: Bool
-  @Published var mcpIntegrationEnabled: Bool
+  var obsidianIntegrationEnabled: Bool {
+    didSet {
+      preferencesStore.set(obsidianIntegrationEnabled, for: .obsidianIntegrationEnabled)
+      onIntegrationStateChanged?()
+    }
+  }
+  var googleCalendarIntegrationEnabled: Bool {
+    didSet {
+      preferencesStore.set(googleCalendarIntegrationEnabled, for: .googleCalendarIntegrationEnabled)
+      onIntegrationStateChanged?()
+    }
+  }
+  var mcpIntegrationEnabled: Bool {
+    didSet {
+      preferencesStore.set(mcpIntegrationEnabled, for: .mcpIntegrationEnabled)
+      onIntegrationStateChanged?()
+    }
+  }
 
   // MARK: - Integration state
 
-  @Published var obsidianInboxPath: String
-  @Published var mcpServerCommandPath: String
-  @Published var pendingObsidianSyncTaskIds: [Int]
-  @Published var googleCalendarEventLinksByTaskKey: [String: String]
+  var obsidianInboxPath: String
+  var mcpServerCommandPath: String
+  var pendingObsidianSyncTaskIds: [Int]
+  var googleCalendarEventLinksByTaskKey: [String: String]
 
   // MARK: - Plugin references
 
@@ -45,7 +61,7 @@ class IntegrationCoordinator: ObservableObject {
 
   // MARK: - Internal state
 
-  var hasPendingSyncProcessingTask = false
+  @ObservationIgnored var hasPendingSyncProcessingTask = false
 
   private static let pendingObsidianSyncDefaultsKey = "pendingObsidianSyncTaskIdsByListId"
   let pendingSyncQueueStore = ListScopedTaskIDStore(
@@ -100,29 +116,6 @@ class IntegrationCoordinator: ObservableObject {
     self.pendingObsidianSyncTaskIds = ListScopedTaskIDStore(
       defaultsKey: IntegrationCoordinator.pendingObsidianSyncDefaultsKey
     ).load(for: initialListId)
-
-    setupBindings()
-  }
-
-  private func setupBindings() {
-    $obsidianIntegrationEnabled
-      .sink { [weak self] value in
-        self?.preferencesStore.set(value, for: .obsidianIntegrationEnabled)
-        self?.onIntegrationStateChanged?()
-      }
-      .store(in: &cancellables)
-    $googleCalendarIntegrationEnabled
-      .sink { [weak self] value in
-        self?.preferencesStore.set(value, for: .googleCalendarIntegrationEnabled)
-        self?.onIntegrationStateChanged?()
-      }
-      .store(in: &cancellables)
-    $mcpIntegrationEnabled
-      .sink { [weak self] in
-        self?.preferencesStore.set($0, for: .mcpIntegrationEnabled)
-        self?.onIntegrationStateChanged?()
-      }
-      .store(in: &cancellables)
   }
 
   // MARK: - Pending Obsidian Sync Queue
@@ -215,12 +208,12 @@ class IntegrationCoordinator: ObservableObject {
 
   // MARK: - Google Calendar
 
-  /// Returns an error message on failure, nil on success.
-  func openTaskInGoogleCalendar(taskId explicitTaskId: Int? = nil) -> String? {
+  func openTaskInGoogleCalendar(taskId explicitTaskId: Int? = nil) {
     guard googleCalendarIntegrationEnabled else {
-      return "Enable Google Calendar integration in Preferences first."
+      onError?("Enable Google Calendar integration in Preferences first.")
+      return
     }
-    guard let ds = dataSource else { return "Internal error: no data source." }
+    guard let ds = dataSource else { onError?("Internal error: no data source."); return }
 
     let selectedTask: CheckvistTask?
     if let explicitTaskId {
@@ -229,7 +222,8 @@ class IntegrationCoordinator: ObservableObject {
       selectedTask = ds.currentTask
     }
     guard let selectedTask else {
-      return "No task selected."
+      onError?("No task selected.")
+      return
     }
 
     let listId = ds.listId
@@ -250,17 +244,15 @@ class IntegrationCoordinator: ObservableObject {
         if let url = outcome.urlToOpen, url.scheme?.lowercased() == "https" {
           NSWorkspace.shared.open(url)
         }
-        // Post error/success via the data source's errorMessage — handled by coordinator
+        self.onError?(nil)
       } catch {
-        // Error handling delegated to coordinator
+        self.onError?(error.localizedDescription)
       }
     }
-
-    return nil  // async — errors posted separately
   }
 
-  /// Async version that returns error message.
-  func openTaskInGoogleCalendarAsync(taskId explicitTaskId: Int? = nil) async -> String? {
+  /// Async version that returns error message (used internally).
+  private func openTaskInGoogleCalendarAsync(taskId explicitTaskId: Int? = nil) async -> String? {
     guard googleCalendarIntegrationEnabled else {
       return "Enable Google Calendar integration in Preferences first."
     }
@@ -308,23 +300,25 @@ class IntegrationCoordinator: ObservableObject {
     }
   }
 
-  /// Returns an error message on failure, nil on success.
-  func openSavedGoogleCalendarEventLink(taskId explicitTaskId: Int? = nil) -> String? {
+  func openSavedGoogleCalendarEventLink(taskId explicitTaskId: Int? = nil) {
     guard googleCalendarIntegrationEnabled else {
-      return "Enable Google Calendar integration in Preferences first."
+      onError?("Enable Google Calendar integration in Preferences first.")
+      return
     }
-    guard let ds = dataSource else { return "Internal error: no data source." }
+    guard let ds = dataSource else { onError?("Internal error: no data source."); return }
     let targetTaskId = explicitTaskId ?? ds.currentTask?.id
     guard let targetTaskId else {
-      return "No task selected."
+      onError?("No task selected.")
+      return
     }
     guard let url = googleCalendarEventLinkURL(taskId: targetTaskId, listId: ds.listId),
       url.scheme?.lowercased() == "https"
     else {
-      return "No saved browser link for this Google Calendar event."
+      onError?("No saved browser link for this Google Calendar event.")
+      return
     }
     NSWorkspace.shared.open(url)
-    return nil
+    onError?(nil)
   }
 
   // MARK: - MCP
@@ -333,12 +327,12 @@ class IntegrationCoordinator: ObservableObject {
     mcpServerCommandPath = mcpIntegrationPlugin.serverCommandURL()?.path ?? ""
   }
 
-  /// Returns an error message on failure, nil on success.
-  func copyMCPClientConfigurationToClipboard() -> String? {
+  func copyMCPClientConfigurationToClipboard() {
     guard mcpIntegrationEnabled else {
-      return "Enable MCP integration in Preferences first."
+      onError?("Enable MCP integration in Preferences first.")
+      return
     }
-    guard let ds = dataSource else { return "Internal error: no data source." }
+    guard let ds = dataSource else { onError?("Internal error: no data source."); return }
 
     refreshMCPServerCommandPath()
 
@@ -351,22 +345,25 @@ class IntegrationCoordinator: ObservableObject {
     _ = NSPasteboard.general.setString(config, forType: .string)
 
     if mcpServerCommandPath.isEmpty {
-      return
+      onError?(
         "MCP config copied with placeholder app path. Set BAR_TASKER_MCP_EXECUTABLE_PATH if your app is outside /Applications."
+      )
+    } else {
+      onError?(nil)
     }
-    return nil
   }
 
-  /// Returns an error message on failure, nil on success.
-  func openMCPServerGuide() -> String? {
+  func openMCPServerGuide() {
     guard mcpIntegrationEnabled else {
-      return "Enable MCP integration in Preferences first."
+      onError?("Enable MCP integration in Preferences first.")
+      return
     }
     guard let guideURL = mcpIntegrationPlugin.guideURL() else {
-      return "MCP guide not found. See docs/mcp-server.md in the repo."
+      onError?("MCP guide not found. See docs/mcp-server.md in the repo.")
+      return
     }
     NSWorkspace.shared.open(guideURL)
-    return nil
+    onError?(nil)
   }
 
   func mcpClientConfigurationPreview(credentials: CheckvistCredentials, listId: String) -> String {
@@ -379,18 +376,19 @@ class IntegrationCoordinator: ObservableObject {
 
   // MARK: - Obsidian
 
-  /// Returns true if a folder was chosen, false otherwise. Sets error message via return value.
   @discardableResult
-  func chooseObsidianInboxFolder() -> (success: Bool, error: String?) {
+  func chooseObsidianInboxFolder() -> Bool {
     do {
       if let selectedPath = try obsidianPlugin.chooseInboxFolder() {
         obsidianInboxPath = selectedPath
         onIntegrationStateChanged?()
-        return (true, nil)
+        onError?(nil)
+        return true
       }
-      return (false, nil)
+      return false
     } catch {
-      return (false, "Failed to save Obsidian folder access.")
+      onError?("Failed to save Obsidian folder access.")
+      return false
     }
   }
 
@@ -400,72 +398,58 @@ class IntegrationCoordinator: ObservableObject {
     onIntegrationStateChanged?()
   }
 
-  /// Returns an error message on failure, nil on success.
-  func linkTaskToObsidianFolder(
-    taskId explicitTaskId: Int? = nil
-  ) -> String? {
+  func linkTaskToObsidianFolder(taskId explicitTaskId: Int? = nil) {
     guard obsidianIntegrationEnabled else {
-      return "Enable Obsidian integration in Preferences first."
+      onError?("Enable Obsidian integration in Preferences first.")
+      return
     }
-    guard let ds = dataSource else { return "Internal error: no data source." }
+    guard let ds = dataSource else { onError?("Internal error: no data source."); return }
     guard
       let task = explicitTaskId.flatMap({ id in ds.tasks.first(where: { $0.id == id }) })
         ?? ds.currentTask
     else {
-      return "No task selected."
+      onError?("No task selected.")
+      return
     }
 
     do {
-      if let linkedPath = try obsidianPlugin.chooseLinkedFolder(
-        forTaskId: task.id,
-        taskContent: task.content
-      ) {
-        _ = linkedPath
-        return nil
-      }
-      return nil
+      _ = try obsidianPlugin.chooseLinkedFolder(forTaskId: task.id, taskContent: task.content)
+      onError?(nil)
     } catch {
-      return "Failed to link Obsidian folder."
+      onError?("Failed to link Obsidian folder.")
     }
   }
 
-  /// Returns an error message on failure, nil on success.
-  func createAndLinkTaskObsidianFolder(
-    taskId explicitTaskId: Int? = nil
-  ) -> String? {
+  func createAndLinkTaskObsidianFolder(taskId explicitTaskId: Int? = nil) {
     guard obsidianIntegrationEnabled else {
-      return "Enable Obsidian integration in Preferences first."
+      onError?("Enable Obsidian integration in Preferences first.")
+      return
     }
-    guard let ds = dataSource else { return "Internal error: no data source." }
+    guard let ds = dataSource else { onError?("Internal error: no data source."); return }
     guard
       let task = explicitTaskId.flatMap({ id in ds.tasks.first(where: { $0.id == id }) })
         ?? ds.currentTask
     else {
-      return "No task selected."
+      onError?("No task selected.")
+      return
     }
 
     do {
-      if let createdPath = try obsidianPlugin.createAndLinkFolder(
-        forTaskId: task.id,
-        taskContent: task.content
-      ) {
-        _ = createdPath
-        return nil
-      }
-      return nil
+      _ = try obsidianPlugin.createAndLinkFolder(forTaskId: task.id, taskContent: task.content)
+      onError?(nil)
     } catch {
-      return "Failed to create and link Obsidian folder."
+      onError?("Failed to create and link Obsidian folder.")
     }
   }
 
-  /// Returns an error message on failure, nil on success.
-  func clearTaskObsidianFolderLink(taskId explicitTaskId: Int? = nil) -> String? {
-    guard let ds = dataSource else { return "Internal error: no data source." }
+  func clearTaskObsidianFolderLink(taskId explicitTaskId: Int? = nil) {
+    guard let ds = dataSource else { onError?("Internal error: no data source."); return }
     guard let targetTaskId = explicitTaskId ?? ds.currentTask?.id else {
-      return "No task selected."
+      onError?("No task selected.")
+      return
     }
     obsidianPlugin.clearLinkedFolder(forTaskId: targetTaskId)
-    return nil
+    onError?(nil)
   }
 
   func hasObsidianFolderLink(taskId: Int) -> Bool {
@@ -495,30 +479,26 @@ class IntegrationCoordinator: ObservableObject {
     return nil
   }
 
-  /// Returns an error message on failure, nil on success.
-  func syncTaskToObsidian(
-    taskId explicitTaskId: Int? = nil,
-    openMode: ObsidianOpenMode
-  ) async -> String? {
+  func syncTaskToObsidian(taskId explicitTaskId: Int? = nil, openMode: ObsidianOpenMode) async {
     guard obsidianIntegrationEnabled else {
-      return "Enable Obsidian integration in Preferences first."
+      onError?("Enable Obsidian integration in Preferences first.")
+      return
     }
-    guard let ds = dataSource else { return "Internal error: no data source." }
+    guard let ds = dataSource else { onError?("Internal error: no data source."); return }
     guard let targetTaskId = explicitTaskId ?? ds.currentTask?.id else {
-      return "No task selected."
+      onError?("No task selected.")
+      return
     }
     guard let task = ds.tasks.first(where: { $0.id == targetTaskId }) ?? ds.currentTask else {
-      return "Task not found."
+      onError?("Task not found.")
+      return
     }
 
     let listId = ds.listId
-    let linkedFolderTaskId = obsidianLinkedFolderAncestorTaskId(
-      for: task, taskList: ds.tasks)
+    let linkedFolderTaskId = obsidianLinkedFolderAncestorTaskId(for: task, taskList: ds.tasks)
     if linkedFolderTaskId == nil && obsidianInboxPath.isEmpty {
-      let result = chooseObsidianInboxFolder()
-      if !result.success {
-        return result.error
-      }
+      let success = chooseObsidianInboxFolder()
+      if !success { return }
     }
 
     do {
@@ -530,13 +510,14 @@ class IntegrationCoordinator: ObservableObject {
         syncDate: Date()
       )
       dequeuePendingObsidianSync(taskId: targetTaskId, listId: listId)
-      return nil
+      onError?(nil)
     } catch {
       enqueuePendingObsidianSync(taskId: targetTaskId, listId: listId)
-      return
+      onError?(
         error.localizedDescription.isEmpty
-        ? "Obsidian sync failed. Added to pending queue."
-        : error.localizedDescription
+          ? "Obsidian sync failed. Added to pending queue."
+          : error.localizedDescription
+      )
     }
   }
