+++ Bar Tasker/Models/CommandSuggestion.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+struct CommandSuggestion {
+  let label: String
+  let command: String
+  let preview: String
+  let keybind: String?
+  let submitImmediately: Bool
+}
