+++ Bar Tasker/Models/OnboardingDialog.swift
@@ -1,0 +1,11 @@
+import Foundation
+
+enum OnboardingDialog: String, CaseIterable, Identifiable {
+  case pluginSelection
+  case checkvist
+  case obsidian
+  case googleCalendar
+  case mcp
+
+  var id: String { rawValue }
+}
