+++ Bar Tasker/Models/QuickAddLocationMode.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+enum QuickAddLocationMode: Int, CaseIterable {
+  case defaultRoot
+  case specificParentTask
+}
