+++ Bar Tasker/Models/QuickEntryMode.swift
@@ -1,0 +1,11 @@
+import Foundation
+
+enum QuickEntryMode {
+  case search
+  case addSibling
+  case addChild
+  case editTask
+  case command
+  case quickAddDefault
+  case quickAddSpecific
+}
