+++ Bar Tasker/Models/UndoableAction.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+enum UndoableAction {
+  case add(taskId: Int)
+  case markDone(taskId: Int)
+  case invalidate(taskId: Int)
+  case update(taskId: Int, oldContent: String, oldDue: String?)
+  case restoreOfflineState(snapshot: OfflineTaskStateSnapshot)
+}
