+++ Bar Tasker/Plugins/Native/MCP/NativeMCPIntegrationPlugin+Settings.swift
@@ -4,15 +4,16 @@ import SwiftUI
 extension NativeMCPIntegrationPlugin: PluginSettingsPageProviding {
   var settingsIconSystemName: String { "link" }
 
-  func makeSettingsView(manager: BarTaskerManager) -> AnyView {
+  func makeSettingsView(manager: BarTaskerCoordinator) -> AnyView {
     AnyView(MCPIntegrationPluginSettingsView(manager: manager))
   }
 }
 
 private struct MCPIntegrationPluginSettingsView: View {
-  @ObservedObject var manager: BarTaskerManager
+  var manager: BarTaskerCoordinator
 
   var body: some View {
+    @Bindable var manager = manager
     Section(header: Text("MCP Plugin")) {
       Toggle("Enable MCP integration", isOn: $manager.integrations.mcpIntegrationEnabled)
 
@@ -31,13 +32,13 @@ private struct MCPIntegrationPluginSettingsView: View {
 
           HStack {
             Button("Refresh Path") {
-              manager.refreshMCPServerCommandPath()
+              manager.integrations.refreshMCPServerCommandPath()
             }
             Button("Copy Client Config") {
-              manager.copyMCPClientConfigurationToClipboard()
+              manager.integrations.copyMCPClientConfigurationToClipboard()
             }
             Button("Open Guide") {
-              manager.openMCPServerGuide()
+              manager.integrations.openMCPServerGuide()
             }
             Spacer()
           }
