+++ Package.swift
@@ -10,14 +10,18 @@ let pluginTargetExcludes = [
   "Bar Tasker/BarTaskerApp.swift",
   "Bar Tasker/BarTaskerCacheState.swift",
   "Bar Tasker/BarTaskerCommandExecutor.swift",
-  "Bar Tasker/BarTaskerManager.swift",
-  "Bar Tasker/BarTaskerManager+Integrations.swift",
-  "Bar Tasker/BarTaskerManager+PreferencesAndShortcuts.swift",
-  "Bar Tasker/BarTaskerManager+ReorderingAndTiming.swift",
-  "Bar Tasker/BarTaskerManager+StateAndLifecycle.swift",
-  "Bar Tasker/BarTaskerManager+TaskOperations.swift",
-  "Bar Tasker/BarTaskerManager+TaskScoping.swift",
-  "Bar Tasker/BarTaskerManager+Types.swift",
+  "Bar Tasker/BarTaskerCoordinator.swift",
+  "Bar Tasker/BarTaskerCoordinator+Navigation.swift",
+  "Bar Tasker/BarTaskerCoordinator+OfflineState.swift",
+  "Bar Tasker/BarTaskerCoordinator+QuickAdd.swift",
+  "Bar Tasker/BarTaskerCoordinator+ReorderingAndTiming.swift",
+  "Bar Tasker/BarTaskerCoordinator+StateAndLifecycle.swift",
+  "Bar Tasker/BarTaskerCoordinator+TaskMutations.swift",
+  "Bar Tasker/BarTaskerCoordinator+TaskScoping.swift",
+  "Bar Tasker/BarTaskerCoordinator+TaskSync.swift",
+  "Bar Tasker/BarTaskerCoordinator+Types.swift",
+  "Bar Tasker/Models",
+  "Bar Tasker/BarTaskerCoordinator+Undo.swift",
   "Bar Tasker/BarTaskerPreferencesStore.swift",
   "Bar Tasker/BarTaskerReorderQueue.swift",
   "Bar Tasker/BarTaskerTaskVisibilityEngine.swift",
